Major API change (#18)
This commit gets rid of most of the API functions. Class methods are added to the sql alchemy models which allow the user to create resources according to constraints needed. The preferred method of working with the datastore should be through the sqlalchemy objects.

- a base Resource class is created.
- The Resource class has class method `permissions_query`. This is a method that can be used to query the database using a permission (read, write, update, etc). Hierarchical considerations are built in (ex - A user has read permission for a job if they have permission for the job or for the project containing the job).
- The Resource class has a get method which allows setting a the desired permission, offset, limit, sorting options (sort by and ascending or descending), specification of column to return if all columns are not desired.
- Objects - Job, Flowchart, Project inherits from Resource. This ensures the same get methods for all resources.
create methods are added to Job, Flowchart, Project. These methods are alternate constructors that ensure that the objects being built adhere to standards necessary for the datastore. Notably, they create, but do not add the objects to the database. For example, the Job.create method checks for the existence of necessary file paths, checks permissions for the projects the to which the job is added, reads job_data.json, and parses the flowchart for the job.
- update methods are added to Job, Project to ensure that updates take place in constraints of datastore as well.